Add SHA-2 support
This patch adds sha-2 support.  The "password" commands gets a new option
--encrypted (generic for all formats, handles MD5, SHA-256, SHA-512).

A new utility, grub-crypt (written in Python) can be used to create the
encrypted passwords; an internal GRUB command corresponding to "md5crypt" is
not provided.

Tested on i386 build (including the #ifdef-ed out built-in tests in the
sha*crypt.c files).  x86_64 build currently fails both on F12 and rawhide,
apparently because there is no glibc-static.i686 available in the x86_64
repositories.

To test:
* Keep existing "password --md5", verify passwords are handled correctly.
* For $format in "--md5", "--sha-256", "--sha-512":
* Use (grub-crypt $format) to generate an encrypted password
* Add "password --encrypted the_generated_password" to grub.conf
* Verify the password is handled correctly.