grub making uefi calls without aligning stack pointer
On a prototype system here at Dell, RHEL6 will not boot in UEFI mode... in
fact, as soon as the UEFI boot manager tries to run grub, the screen goes black
with a cursor in the middle of the screen.

It appears that grub is making calls to the UEFI firmware when the stack
pointer is not aligned to a 16-byte boundary (which is required by the UEFI 2.1
spec).  (I enlisted the aid of a firmware programmer with an xdp on the system
to help figure this out...)

I recompiled grub, changing out the #defines for Call_Service*() (in
/efi/grub/efi/eficall.h) to use uefi_call_wrapper() (from gnu-efi) instead of
x64_call*() (from grub's efi/x86_64/callwrap.c), and grub booted up and
presented me a menu.